+++ packages/tx/src/generic/generic/makeEmptyArrayIfUndefined.ts
@@ -1,3 +1,5 @@
+import type {Maybe} from '@datakit/types';
+
 /**
  * Return an empty array if the input is undefined, identity otherwise.
  *
@@ -9,5 +11,5 @@
  *
  * @since 0.1.0
  */
-export const makeEmptyArrayIfUndefined = <T>(x: T | undefined): T | never[] =>
+export const makeEmptyArrayIfUndefined = <T>(x: Maybe<T>): T | never[] =>
 	x === undefined ? [] : x;