+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export type * from './array';
 export type * from './function';
+export type * from './nil';
 export type * from './object';
-export type * from './shared';
+export type * from './shape';