+++ packages/types/src/nil.ts
@@ -1,0 +1,14 @@
+/**
+ * A value that can be of type T or undefined.
+ */
+export type Maybe<T> = T | undefined;
+
+/**
+ * A value that can be of type T or null.
+ */
+export type Nullable<T> = T | null;
+
+/**
+ * A value that can be of type T, null, or undefined.
+ */
+export type Nilable<T> = T | null | undefined;