+++ packages/tx/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @datakit/tx
 
+## next
+
+- updated `@data/types`, used `Maybe`
+
 ## v0.3.0
 
 Added: